feat : add noSearchData & migrate main to index
