+++ src/components/mainContent/ListSearchSection/NoSearchData.tsx
@@ -1,0 +1,12 @@
+import noData from '@/public/assets/icon/NoData.png';
+import Image from 'next/image';
+
+const NoSearchData = () => {
+  return (
+    <div className="mt-80 text-20 font-bold">
+      <p>검색 결과가 없습니다.</p>
+    </div>
+  );
+};
+
+export default NoSearchData;
