+++ src/pages/index.tsx
@@ -1,11 +1,20 @@
 import Layout from '../components/common/layout';
+import CarouselSection from '@/src/components/mainContent/CarouselSection';
+import ListSearchSection from '@/src/components/mainContent/ListSearchSection';
 
-const Home = () => {
+const main = () => {
   return (
     <Layout>
-      <h1>Home</h1>
+      <div className="flex flex-col items-center p-100">
+        <header>
+          <CarouselSection />
+        </header>
+        <main>
+          <ListSearchSection />
+        </main>
+      </div>
     </Layout>
   );
 };
 
-export default Home;
+export default main;
